requirements.txt
fastapi
uvicorn
pydantic
requests
groq
python-dotenv
httpx
